// src/types.ts
export type PokemonType = {
  name: string;
  url: string;
};

export type ItemType = {
  abilities: any[];
  base_experience: number;
  forms: [];
  game_indices: [];
  height: number;
  held_items: [];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: [];
  name: string;
  order: number;
  past_types: [];
  species: {};
  sprites: any;
  stats: [];
  types: [];
  weight: number;
  image: string;
};